Agregada la clase Operacion con controlador y servicio

+++ src/main/java/com/challenge/repositorio/TarjetaRepositorio.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.challenge.model.Tarjeta;
 
-public interface TarjetaRepositorio extends JpaRepository<Tarjeta, Long> {
+public interface TarjetaRepositorio extends JpaRepository<Tarjeta, Integer> {
 
 }
